Move main logic to handler
